Agent/requirements.txt
fastapi
uvicorn[standard]
python-dotenv
PyPDF2
pillow
pytesseract
scikit-learn
numpy
scipy
langchain-google-genai
requests
transformers
torch